requirements.txt
opencv-python
pathlib
requests
pillow
numpy 
picamera2